+++ cli/lib/src/functions/functions.dart
@@ -5,3 +5,5 @@ export 'add/add.dart';
 export 'components_path/components_path.dart';
 export 'close.dart';
 export 'component_methods/component_methods.dart';
+export 'help/help_command.dart';
+export 'version/version_command.dart';