+++ cli/lib/src/functions/help/help_command.dart
@@ -1,0 +1,11 @@
+void helpCommand() {
+  print('''
+      init
+        - Initialize the project
+        
+      add <componentName> 
+        - Add a new component
+      
+      [Visit https://fcnui.shoh.dev/docs/get_started for more information]
+      ''');
+}