+++ cli/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.4 (15.02.2024)
+
+- Add help command - `fcnui help`. Prints all available commands.
+- Add version command - `fcnui version`. Prints the current version of the cli package.
+- [Fix] `fcnui.json` is created if inside a Flutter project.
+
 ## 1.0.3
 
 - Add debug mode switcher in constants.dart.